add a display value to the Param component
this will be the normalised numerical value or dropdown label
