+++ src/Param.svelte
@@ -1,20 +1,9 @@
 <script>
+  import { formatValue } from "./lib/utils.js"
+
   export let param = null
   
-  function normalise(x) {
-    return (paramMax - paramMin) * ((x-xMin)/(xMax - xMin)) + paramMin;
-  }
-  
-  function round(num) {
-    return Math.round( num * 10 + Number.EPSILON ) / 10;
-  }
-  
-  const paramMin = -100;
-  const paramMax = 100;
-  const xMin = -1024;
-  const xMax = 1024;
-  
-  let val = param;
+  let val = param.value;
   
   let rangeMin = param.min < 0 ? -1024 : 0;
   let rangeMax = 1024;
@@ -46,6 +35,9 @@
     <h3>
       {param.name}
     </h3>
+    <p>
+      {param.displayValue}
+    </p>
   </header>
 <!--   <code>
     id: {param.id}<br>
@@ -77,37 +69,37 @@
     <table>
       <tr>
         <td>Base</td>
-        <td>{round(normalise(param.value))}%<br></td>
+        <td>{formatValue(param.value)}%<br></td>
       </tr>
       <tr>
         <td>Env</td>
-        <td>{round(normalise(param.mods.env))}%<br></td>
+        <td>{formatValue(param.mods.env)}%<br></td>
       </tr>
       <tr>
         <td>Pressure</td>
-        <td>{round(normalise(param.mods.pressure))}%<br></td>
+        <td>{formatValue(param.mods.pressure)}%<br></td>
       </tr>
       <tr>
         <td>A</td>
-        <td>{round(normalise(param.mods.a))}%<br></td>
+        <td>{formatValue(param.mods.a)}%<br></td>
       </tr>
     </table>
     <table>
       <tr>
         <td>B</td>
-        <td>{round(normalise(param.mods.b))}%<br></td>
+        <td>{formatValue(param.mods.b)}%<br></td>
       </tr>
       <tr>
         <td>X</td>
-        <td>{round(normalise(param.mods.x))}%<br></td>
+        <td>{formatValue(param.mods.x)}%<br></td>
       </tr>
       <tr>
         <td>Y</td>
-        <td>{round(normalise(param.mods.y))}%<br></td>
+        <td>{formatValue(param.mods.y)}%<br></td>
       </tr>
       <tr>
         <td>Random</td>
-        <td>{round(normalise(param.mods.random))}%<br></td>
+        <td>{formatValue(param.mods.random)}%<br></td>
       </tr>
     </table>
   </div>
